Add unsupported system error if not linux/windows/mac
I was getting type errors for the install_fixes file saying that there
was a possibility that site_packages_dir and build_dir were undefined.

It may be preferable to assume that anything that isn't windows or mac
follows the convention for linux, but for the moment we can raise an
exception.