Attempting to change the code to run on MPS (Failing)
Changing the implementation to run on an M1 Apple device.

Easy changes:
- convert calls to `.cuda()` to `.to(device)`
- convert calls to `torch.load(f)` to `torch.load(f, device)`

We are not able to get the memory from the MPS device, so just stuck an
arbitrary amount in for the moment.

Ran into issue with MPS not supporting Float64. Convert to Float32
caused further issue with type mismatch:

```
Input type (torch.FloatTensor) and weight type (MPSFloatType) should be the same
```

We tried to change how the tensor is converted to float, but weren't able to pin
the problem down.